Rename code_2 to code 2
Describing a basic iterative optimization algorithm with a gradient estimator


At each round t, our CODE 2 first collects a sample of size n(t) from current distribution Pt = (1-e)D(x(t))+ eQ, then applies code 1 as a sub-routine to compute ~g(t) as an estimation of current gradient mean, after that conducts gradient descent using the estimated gradient, and finally deploies the updated model x(t+1) to the environment.



SOME DESCRIPTIONS :

The NumPy zeros() function in Python is used to create an array of specified shapes and types, with all elements initialized to zero and similarly np.ones() creates an array with elements initialized to one.Data type of the returned array is float by default and in our program, np.ones() with dtype=int will return array with ones of integer type.np.ones(T,dtype=int)*10 will create an array with 'T' number of tens of integer type.

np.array creates an array with any number of dimensions and when we use np.sum() with axis = 0, the function will sum over the 0th axis (the rows). It’s basically summing up the values row-wise, and producing a new array (with lower dimensions).


The numpy.random.randn() function creates an array of specified shape and fills it with random values as per standard normal distribution.


Please note that one needs to replace the placeholder data gradient_sample_S with his/her actual gradient sample data. Also, the implementation of the GRADIENTESTIMATOR and MEAN functions is missing and should be replaced with the specific implementation