Rename code1 to code 1
the Robust Gradient Estimator

SOME DESCRIPTIONS

SciPy : A python library which SciPy provides algorithms for optimization, integration, interpolation, eigenvalue problems, algebraic equations, differential equations, statistics and many other classes of problems.

KD Tree is a space-partitioning data structure for organizing points in a k-dimensional space,i.e., each data point has k properties.It is used for range searching in higher dimension(for example, in 1D range search we build a BST and split values in 1dimension).The idea is that each level of the tree compares against 1 dimension.scipy.spatial.cKDTree class provides an index into a set of k-dimensional points which can be used to rapidly look up the nearest neighbors(balls) of any point.

Scikit-Learn, also known as sklearn is a python library to implement machine learning models and statistical modellings.sklearn.decomposition.PCA can be utilized for extracting information from a high-dimensional space by projecting it into a lower-dimensional sub-space. It tries to preserve the essential parts that have more variation of the data and remove the non-essential parts with fewer variatio.PCA stands for Principal Component Analysis.

Confidence intervals->refers to the probability that a population parameter will fall between a set of values for a certain proportion of times.

The query() method allows you to query the DataFrame. The query() method takes a query expression as a string parameter, which has to evaluate to either True of False. It returns the DataFrame where the result is True according to the query expression.

Please note that one needs to replace the placeholder data gradient_sample_S with his/her actual gradient sample data. Also, the implementation of the GRADIENTESTIMATOR and MEAN functions is missing and should be replaced with the specific implementation